Fixed log bugs, leak bugs, found spent. fire bug
Bug in neuron.c, need to verify that the neuronID is columnar and not
pointing at garbage memory.
Fixed some neuron log issues too
Still need to make the neuron event log function more robust.

+++ assist.h
@@ -45,7 +45,7 @@ struct NeuronModel;
 
 
 //Proper Typdefs:
-typedef uint64_t id_type;
+typedef uint16_t id_type;
 typedef int64_t volt_type;
 typedef int64_t weight_type;
 typedef uint_fast32_t thresh_type;
